Power Cells, Lasers and Pushable Blocks
Features
+ Power cell objects can be picked up in order to power the level exit.
+ Basic Wall Lasers have been added.  They currently have no
functionality but the distance of the laser is stopped by solid objects.
+ A script for making objects pushable has been added.  Currently only
the player can move things.  Enemies will probably just pass through
them.

+++ Assets/Scripts/UI.cs
@@ -25,7 +25,7 @@ public class UI : MonoBehaviour {
 	// Update is called once per frame
 	void Update () {
         lives.text = "" + gm.playerLives;
-        collected.fillAmount = p.curPower / p.maxPower;
+        collected.fillAmount = p.curPower / (float)p.maxPower;
         time.text = gm.gameTime.GetTime();
         yellowKey.gameObject.SetActive(p.HasKey(Key.Type.Yellow));
         redKey.gameObject.SetActive(p.HasKey(Key.Type.Red));
